chore(shared): remove miscelaneous types from variables
