+++ src/Components/Forms/SelectFieldButton.tsx
@@ -10,7 +10,12 @@ import {
 import { Button } from '@grafana/ui';
 import { VariableHide } from '@grafana/schema';
 import { addToFavoriteServicesInStorage } from 'services/store';
-import { VAR_DATASOURCE, StartingPointSelectedEvent } from 'services/shared';
+import { VAR_DATASOURCE } from 'services/shared';
+import { BusEventBase } from '@grafana/data';
+
+export class StartingPointSelectedEvent extends BusEventBase {
+  public static type = 'start-point-selected-event';
+}
 
 export interface SelectFieldButtonState extends SceneObjectState {
   value: string;
